Må bruke refresh for å få en komplett liste over alle prosjekter i et program
