+++ SharePointFramework/ProgramWebParts/src/components/ProgramAdministration/AddProjectDialog/useAddProjectDialog.ts
@@ -1,6 +1,5 @@
 import { useContext, useEffect } from 'react'
 import { ProgramAdministrationContext } from '../context'
-import { getHubSiteProjects } from '../data'
 import { DATA_LOADED, SET_SELECTED_TO_ADD } from '../reducer'
 import { useRowRenderer } from '../useRowRenderer'
 import { useSelectionList } from '../useSelectionList'
@@ -14,7 +13,7 @@ export const useAddProjectDialog = () => {
   })
 
   useEffect(() => {
-    getHubSiteProjects()
+    context.props.dataAdapter.getHubSiteProjects()
       .then((availableProjects) =>
         context.dispatch(DATA_LOADED({ data: { availableProjects }, scope: 'AddProjectDialog' }))
       )
